Fix clippy warnings (batch 1-4)

+++ src/perception/hologram_decoder.rs
@@ -40,7 +40,7 @@ impl HologramDecoder {
             let center_x = manifold.centers_x[e].round() as i32;
             let center_y = manifold.centers_y[e].round() as i32;
 
-            if center_x >= 0 && center_x < width as i32 && center_y >= 0 && center_y < height as i32
+            if center_x >= 0 && (center_x as usize) < width && center_y >= 0 && (center_y as usize) < height
             {
                 // Untuk Swarm Kuantum dasar, token warna sudah tersimpan utuh.
                 // Jika ingin ekstraksi Fasa Semantik bisa panggil Probe Warna.
